Rename common cached pipeline seam
